We are live (repo moved public, website operational). Additional edits to the documentation can now take place.
Updated Coereba_GateCutoffs to use openCyto gate_mindensity for the initial gates, which speeds up vs. the old implementation. We still have clear misses, but can work at integrating in the decision tree checks of the old one when there is time.

Additionally, setting up a way to skip this by provisioning a fluorophore gating template with more stringent gate args for customization.